chore: Removes all (right now) unnecessary code, strips down contracts to a bare sensible minimum
