--- test/helpers/latestTime.js
@@ -1,4 +1,0 @@
-// Returns the time of the last mined block in seconds
-module.exports = function latestTime () {
-  return web3.eth.getBlock('latest').timestamp;
-};
